# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Extractor",
    "description": "Brazilian economic indicators extraction, storage, and dashboard API. 25 macro indicators from BCB SGS, IBGE SIDRA, and B3/Yahoo Finance. Supports on-demand transforms, pinned dashboards, and release calendars.",
    "version": "0.5.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/series": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "List all economic indicator series",
        "description": "Returns all 25 configured series with their metadata and current extraction status. Optionally filter by category (e.g. 'Inflação', 'Juros', 'Câmbio', 'Mercado', 'Atividade', 'Trabalho', 'Fiscal', 'Externo').",
        "operationId": "list_series_series_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category name (case-sensitive, pt-BR). Omit to return all.",
              "title": "Category"
            },
            "description": "Filter by category name (case-sensitive, pt-BR). Omit to return all."
          }
        ],
        "responses": {
          "200": {
            "description": "Series list returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/series/{code}": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Get single series metadata",
        "description": "Returns full metadata for one economic indicator series, including source information, frequency, unit, and current extraction status. Used by the Metadados dossier page.",
        "operationId": "get_series_series__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Series metadata returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesRead"
                }
              }
            }
          },
          "404": {
            "description": "Series code not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/series/{code}/observations": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Get raw observations for a series",
        "description": "Returns stored time-series observations for the given series. Supports optional date range filtering via `from` and `to` (ISO-8601). Results are ordered by observed_at ascending. Maximum 5000 rows returned per request (default 500). Decimal values stored in Postgres are returned as float64.",
        "operationId": "get_observations_series__code__observations_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (inclusive). ISO-8601: YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS.",
              "title": "From"
            },
            "description": "Start date (inclusive). ISO-8601: YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (inclusive). ISO-8601: YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS.",
              "title": "To"
            },
            "description": "End date (inclusive). ISO-8601: YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "description": "Maximum rows to return. Default 500, max 5000.",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum rows to return. Default 500, max 5000."
          }
        ],
        "responses": {
          "200": {
            "description": "Observations returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Series code not found."
          },
          "422": {
            "description": "Invalid date format or limit out of range."
          }
        }
      }
    },
...
    "/admin/extract/{code}": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Manually trigger extraction for a series",
        "description": "Triggers an immediate extraction run for the given series code. Fetches new observations from the configured upstream source (BCB SGS, IBGE SIDRA, or Yahoo Finance/B3), upserts them into the database, and updates series.status and last_success_at. Incremental: fetches since last_success_at if set, else full backfill. Returns outcome with count of inserted/updated observations. Note: No authentication required in v1 (single-user local deployment). LIVE NETWORK CALL — runs against real upstream APIs.",
        "operationId": "trigger_extraction_admin_extract__code__post",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Extraction completed (check status field for success/failure).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionResultResponse"
                }
              }
            }
          },
          "404": {
            "description": "Series code not found."
          },
          "422": {
            "description": "Series source cannot be mapped to a known adapter."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExtractionResultResponse": {
        "properties": {
          "series_code": {
            "type": "string",
            "title": "Series Code",
            "description": "Series code that was extracted."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Outcome: 'success' | 'failed'. Mirrors series.status after the run."
          },
          "observations_upserted": {
            "type": "integer",
            "title": "Observations Upserted",
            "description": "Number of observation rows inserted or updated."
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At",
            "description": "Timestamp of the most recent observation after this extraction. Null on failure."
          },
          "extraction_at": {
            "type": "string",
            "format": "date-time",
            "title": "Extraction At",
            "description": "UTC timestamp when the extraction was triggered."
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if the extraction failed. Null on success."
          }
        },
        "type": "object",
        "required": [
          "series_code",
          "status",
          "observations_upserted",
          "extraction_at"
        ],
        "title": "ExtractionResultResponse",
        "description": "Response schema for POST /admin/extract/{code}.\n\nReports the outcome of a manual extraction trigger."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObservationListResponse": {
        "properties": {
          "series_code": {
            "type": "string",
            "title": "Series Code",
            "description": "Series code these observations belong to."
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ObservationRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Observations ordered by observed_at ascending."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total rows matching the query before limit was applied."
          },
          "returned": {
            "type": "integer",
            "title": "Returned",
            "description": "Number of rows in this response."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Effective limit applied (max 5000, default 500)."
          },
          "from_dt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Dt",
            "description": "Lower bound applied (inclusive), if provided."
          },
          "to_dt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Dt",
            "description": "Upper bound applied (inclusive), if provided."
          }
        },
        "type": "object",
        "required": [
          "series_code",
          "items",
          "total",
          "returned",
          "limit"
        ],
        "title": "ObservationListResponse",
        "description": "Response schema for GET /series/{code}/observations."
      },
      "ObservationRead": {
        "properties": {
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At",
            "description": "UTC timestamp of the observation (timezone-aware ISO-8601)."
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Numeric value as float64. Source precision: Numeric(20, 6)."
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At",
            "description": "UTC timestamp when this row was first written to the database."
          }
        },
        "type": "object",
        "required": [
          "observed_at",
          "value",
          "ingested_at"
        ],
        "title": "ObservationRead",
        "description": "A single time-series data point.\n\nNote: `value` is returned as float (converted from Decimal at the DB\nboundary). Precision is preserved to 6 decimal places by the DB schema."
      },
...
      "SeriesListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SeriesRead"
            },
            "type": "array",
            "title": "Items",
            "description": "List of all series ordered by category then code."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of series returned."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SeriesListResponse",
        "description": "Response schema for GET /series."
      },
      "SeriesRead": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Series primary key, e.g. 'IPCA'."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable display name (pt-BR)."
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Grouping category, e.g. 'Inflação', 'Juros'."
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Data source name, e.g. 'BCB SGS', 'IBGE SIDRA'."
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "description": "Source-system identifier, e.g. '433' for BCB SGS."
          },
          "frequency": {
            "type": "string",
            "title": "Frequency",
            "description": "Release cadence: 'daily' | 'monthly' | 'quarterly' | 'event'."
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Unit of measure, e.g. '%', 'R$', 'pts'."
          },
          "first_observation": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Observation",
            "description": "Date of the earliest known observation for this series."
          },
          "last_extraction_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Extraction At",
            "description": "UTC timestamp of the most recent extraction attempt."
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At",
            "description": "UTC timestamp of the most recent successful extraction."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Freshness status: 'fresh' | 'stale' | 'failed'."
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Source-specific metadata (methodology notes, URLs, etc.)."
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "category",
          "source",
          "source_id",
          "frequency",
          "unit",
          "status"
        ],
        "title": "SeriesRead",
        "description": "Full metadata for a single economic indicator series.\n\nNumeric values stored as Decimal in Postgres are returned as float here."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}